search_config: Do not match subsets of words when searching for a term
The regular expressions created from search terms did not anchor to
the start of a line, a leading space, a trailing space, or the end of a
line. As a result, false positive matches would occur if a provided term
was a subset of a word. (For example an occurrence "board" would
be incorrectly reported if the word "motherboard" was encountered).

These missing anchors have been added to correct this.