reid/convert: Call pdftotext and gosseract directly
It doesn't appear to be possible to use docconv's ConvertPDFText and
ConvertImages directly because the resulting text is tucked away in
an unexported `body` field. Therefore, it doesn't seem possible
to selectively utilize OCR with docconv; it's either enabled and in use
or not enabled.

Dropping docconv dependency; I was really only scratching the surface
of its use-cases anywany...