Added short minifile warning and forced OCR conversion
We've been running into some issues where a coverpage or copyright
notice contains searchable text, but the actual content is contained
in scanned images.

reid now warns about when it finds a minified document with a low
character-count.

A new boolean flag has been added to reid.Convert, and accordingly,
reid-convert to force the use of OCR.